requirements.txt
numpy==1.17.4
scipy==1.3.1
# torchvision==0.4.1
pandas==0.25.2
matplotlib==3.1.1
pytorch-lightning==1.3.8
h5py==2.10.0
Pillow==7.1.2
albumentations==0.4.5
pytest
